pyproject.toml
[build-system]
requires = ["setuptools>=75.6.0", "wheel"]
build-backend = "setuptools.build_meta"
[project]
name = "PalworldSaveTools"
version = "1.1.31"
description = "All-in-one tool for fixing/transferring/editing Palworld saves"
readme = "readme.md"
license = "MIT"
authors = [{ name = "Pylar" }]
requires-python = ">=3.11"
dependencies = [
    "loguru>=0.7.3",
    "matplotlib",
    "pandas",
    "customtkinter",
    "cityhash",
    "pillow",
    "numpy",
    "pygame",
    "pyside6-essentials",
    "pyooz @ git+https://github.com/oMaN-Rod/pyooz.git",
    "cx-freeze>=8.3.0",
    "nerdfont"
]
[tool.setuptools.packages.find]
where = ["."]
include = ["Assets*"]